Rotating Square Rose
This is a simple square the reduces in size toward the center. Each square spins at a slightly faster speed causin a hypnotic effect. Also the color of each square gradually reduces from white to black giving the appearance of pedals. The center square and background grid are controlled by the amplitude.